Refactor: 구매자 관련 엔티티 패키지 이동 및 ID 컬럼명 정규화
