+++ src/main/java/backend/mossy/boundedContext/cash/domain/core/PayMethod.java
@@ -1,4 +1,4 @@
-package backend.mossy.boundedContext.cash.domain.payment;
+package backend.mossy.boundedContext.cash.domain.core;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
