+++ src/main/java/backend/mossy/boundedContext/cash/domain/core/PaymentStatus.java
@@ -1,4 +1,4 @@
-package backend.mossy.boundedContext.cash.domain.payment;
+package backend.mossy.boundedContext.cash.domain.core;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
