+++ src/main/java/backend/mossy/boundedContext/cash/domain/user/UserCashLog.java
@@ -1,9 +1,7 @@
-package backend.mossy.boundedContext.cash.domain.history;
+package backend.mossy.boundedContext.cash.domain.user;
 
 import static jakarta.persistence.FetchType.LAZY;
 
-import backend.mossy.boundedContext.cash.domain.wallet.CashUser;
-import backend.mossy.boundedContext.cash.domain.wallet.Wallet;
 import backend.mossy.global.jpa.entity.BaseIdAndTime;
 import jakarta.persistence.*;
 import java.math.BigDecimal;
@@ -12,15 +10,15 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Entity
-@Table(name = "CASH_CASH_LOG")
+@Table(name = "CASH_USER_LOG")
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AttributeOverride(name = "id", column = @Column(name = "cash_log_id"))
-public class CashLog extends BaseIdAndTime {
+public class UserCashLog extends BaseIdAndTime {
 
     @Enumerated(EnumType.STRING)
     @Column(nullable = false)
-    private EventType eventType;
+    private UserEventType eventType;
     @Column(nullable = false)
     private String relTypeCode;
     @Column(nullable = false)
@@ -29,8 +27,8 @@ public class CashLog extends BaseIdAndTime {
     @JoinColumn(name = "user_id", nullable = false, foreignKey = @ForeignKey(ConstraintMode.NO_CONSTRAINT))
     private CashUser user;
     @ManyToOne(fetch = LAZY)
-    @JoinColumn(name = "wallet_id", nullable = false, foreignKey = @ForeignKey(ConstraintMode.NO_CONSTRAINT))
-    private Wallet wallet;
+    @JoinColumn(name = "user_wallet_id", nullable = false, foreignKey = @ForeignKey(ConstraintMode.NO_CONSTRAINT))
+    private UserWallet wallet;
     @Column(precision = 10, scale = 3, nullable = false)
     private BigDecimal amount = BigDecimal.ZERO;
     @Column(precision = 10, scale = 3, nullable = false)
