+++ src/main/java/backend/mossy/boundedContext/cash/domain/user/UserEventType.java
@@ -1,0 +1,12 @@
+package backend.mossy.boundedContext.cash.domain.user;
+
+public enum UserEventType {
+
+    /* ----------------------------------------------------------------
+     * 구매자(User) 액션 관련 이벤트
+     * ---------------------------------------------------------------- */
+    충전__무통장입금,
+    충전__PG결제_토스페이먼츠,
+    사용__주문결제,       // 구매자가 결제 버튼을 눌러 잔액이 차감될 때
+    환불__결제취소        // 결제 취소로 인해 잔액이 다시 들어올 때
+}
