+++ src/main/java/backend/mossy/boundedContext/cash/domain/user/UserWallet.java
@@ -1,6 +1,5 @@
-package backend.mossy.boundedContext.cash.domain.wallet;
+package backend.mossy.boundedContext.cash.domain.user;
 
-import backend.mossy.boundedContext.cash.domain.history.CashLog;
 import backend.mossy.global.jpa.entity.BaseManualIdAndTime;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
@@ -14,13 +13,13 @@ import static jakarta.persistence.CascadeType.PERSIST;
 import static jakarta.persistence.CascadeType.REMOVE;
 
 @Entity
-@Table(name = "CASH_WALLET")
+@Table(name = "CASH_USER_WALLET")
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Getter
-@AttributeOverride(name = "id", column = @Column(name = "wallet_id"))
-public class Wallet extends BaseManualIdAndTime {
+@AttributeOverride(name = "id", column = @Column(name = "user_wallet_id"))
+public class UserWallet extends BaseManualIdAndTime {
 
-    @ManyToOne(fetch = FetchType.LAZY)
+    @OneToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "user_id", nullable = false, foreignKey = @ForeignKey(ConstraintMode.NO_CONSTRAINT))
     private CashUser user;
 
@@ -28,11 +27,11 @@ public class Wallet extends BaseManualIdAndTime {
     private BigDecimal balance = BigDecimal.ZERO;
 
     @OneToMany(mappedBy = "wallet", cascade = {PERSIST, REMOVE}, orphanRemoval = true)
-    private List<CashLog> cashLogs = new ArrayList<>();
+    private List<UserCashLog> cashLogs = new ArrayList<>();
 
-    public Wallet(CashUser user) {
+    public UserWallet(CashUser user) {
         super(user.getId());
-        this.user = user; // 필드명 변경에 따른 할당 수정
-        this.balance = BigDecimal.ZERO; // 초기값 설정 변경
+        this.user = user;
+        this.balance = BigDecimal.ZERO;
     }
 }
